+++ HISTORY.rst
@@ -25,6 +25,7 @@ Historia zmian
 * raport slotów - ewaluacja pozwala na tworzenie raportów później niż dla
   2021 roku (#1210),
 * definiowalna ilość wyświetlanych jednostek na stronę (#1211),
+* możliwość ukrycia jednostek podrzędnych na stronie prezentacji danych (#1212),
 
 202202.1085
 -----------